requirements.txt
pyaml
requests
beautifulsoup4
stem